Convert bongo to new galaxian driver.
