+++ src/mame/includes/galaxian.h
@@ -258,6 +258,7 @@ public:
 	DECLARE_DRIVER_INIT(moonwar);
 	DECLARE_DRIVER_INIT(ghostmun);
 	DECLARE_DRIVER_INIT(froggrs);
+	DECLARE_DRIVER_INIT(bongo);
 	TILE_GET_INFO_MEMBER(bg_get_tile_info);
 	virtual void video_start();
 	DECLARE_PALETTE_INIT(galaxian);
