+++ src/mame/video/galaxold.c
@@ -756,15 +756,6 @@ VIDEO_START_MEMBER(galaxold_state,ozon1)
 	m_bg_tilemap->set_scrolldx(0, 384-256);
 }
 
-VIDEO_START_MEMBER(galaxold_state,bongo)
-{
-	VIDEO_START_CALL_MEMBER(galaxold_plain);
-
-	m_bg_tilemap->set_scrolldx(0, 384-256);
-
-	m_modify_spritecode = &galaxold_state::batman2_modify_spritecode;
-}
-
 TILE_GET_INFO_MEMBER(galaxold_state::dambustr_get_tile_info2)
 {
 	UINT8 x = tile_index & 0x1f;
